+++ react/components/PONumber.tsx
@@ -3,9 +3,10 @@ import { useIntl } from 'react-intl'
 import { Input } from 'vtex.styleguide'
 
 import { useOrderFormCustom } from '../hooks'
+import type { CompleteOrderForm } from '../typings'
 import {
-  B2B_CUSTOM_APP_ID,
-  B2B_CUSTOM_APP_MAJOR,
+  B2B_CHECKOUT_CUSTOM_APP_ID,
+  B2B_CHECKOUT_CUSTOM_APP_MAJOR,
   messages,
   PO_NUMBER_CUSTOM_FIELD,
 } from '../utils'
@@ -24,15 +25,15 @@ export function PONumber() {
             ...orderForm.customData,
             customApps: [
               {
-                id: B2B_CUSTOM_APP_ID,
-                major: B2B_CUSTOM_APP_MAJOR,
+                id: B2B_CHECKOUT_CUSTOM_APP_ID,
+                major: B2B_CHECKOUT_CUSTOM_APP_MAJOR,
                 fields: {
                   [PO_NUMBER_CUSTOM_FIELD]: e.target.value,
                 },
               },
             ],
           },
-    })
+    } as CompleteOrderForm)
   }
 
   return (