+++ react/components/PaymentData.tsx
@@ -1,18 +1,19 @@
 import React, { useCallback, useEffect } from 'react'
 import { useMutation } from 'react-apollo'
 import { useIntl } from 'react-intl'
-import type { OrderForm, PaymentDataInput } from 'vtex.checkout-graphql'
+import type { PaymentDataInput } from 'vtex.checkout-graphql'
 import type { UpdateOrderFormPaymentMutation } from 'vtex.checkout-resources'
 import { MutationUpdateOrderFormPayment } from 'vtex.checkout-resources'
-import { Dropdown, withToast } from 'vtex.styleguide'
+import { Dropdown } from 'vtex.styleguide'
 
 import { useCheckoutB2BContext } from '../CheckoutB2BContext'
-import { useOrderFormCustom } from '../hooks'
-import type { WithToast } from '../typings'
+import { useOrderFormCustom, useToast } from '../hooks'
+import type { CompleteOrderForm } from '../typings'
 import { getFirstInstallmentByPaymentSystem, messages } from '../utils'
 import { TotalizerSpinner } from './TotalizerSpinner'
 
-function PaymentDataWrapper({ showToast }: WithToast) {
+export function PaymentData() {
+  const showToast = useToast()
   const { formatMessage } = useIntl()
   const { setPending } = useCheckoutB2BContext()
   const {
@@ -26,10 +27,13 @@ function PaymentDataWrapper({ showToast }: WithToast) {
     { paymentData: PaymentDataInput }
   >(MutationUpdateOrderFormPayment, {
     onCompleted({ updateOrderFormPayment }) {
-      setOrderForm(updateOrderFormPayment as OrderForm)
+      setOrderForm({
+        ...orderForm,
+        ...updateOrderFormPayment,
+      } as CompleteOrderForm)
     },
     onError({ message }) {
-      showToast?.({ message })
+      showToast({ message })
     },
   })
 
@@ -115,5 +119,3 @@ function PaymentDataWrapper({ showToast }: WithToast) {
     />
   )
 }
-
-export const PaymentData = withToast(PaymentDataWrapper)