+++ react/hooks/useOrganization.ts
@@ -3,16 +3,18 @@ import { useQuery } from 'react-apollo'
 import type { QueryGetUsersArgs } from 'vtex.b2b-organizations-graphql'
 import type { SessionSuccess } from 'vtex.session-client'
 import { useFullSession } from 'vtex.session-client'
+import type { Query } from 'vtex.storefront-permissions'
 
 import GET_ORGANIZATION from '../graphql/getOrganization.graphql'
+import GET_PERMISSIONS from '../graphql/getPermissions.graphql'
 import type {
   CustomOrganization,
   GetOrganizationQuery,
   SessionOrganizationData,
 } from '../typings'
+import { NAMESPACE_ITEMS, SESSION_NAMESPACE } from '../utils'
 
-const SESSION_NAMESPACE = 'storefront-permissions'
-const NAMESPACE_ITEMS = ['organization', 'costcenter']
+type GetPermissionsQuery = Pick<Query, 'checkUserPermission'>
 
 export function useOrganization() {
   const { data: sessionData, loading: sessionLoading } = useFullSession({
@@ -22,7 +24,7 @@ export function useOrganization() {
   })
 
   const session = sessionData?.session as SessionSuccess | undefined
-  const organizationData: SessionOrganizationData =
+  const organizationData: SessionOrganizationData | undefined =
     session?.namespaces[SESSION_NAMESPACE]
 
   const organizationId = organizationData?.organization.value
@@ -37,14 +39,27 @@ export function useOrganization() {
     variables: { costCenterId, organizationId },
   })
 
+  const {
+    data: permissionsData,
+    loading: permissionsLoading,
+  } = useQuery<GetPermissionsQuery>(GET_PERMISSIONS, {
+    ssr: false,
+  })
+
   const organization: CustomOrganization = useMemo(() => {
     return {
       ...data?.getOrganizationByIdStorefront,
       users: data?.getUsers,
+      costCenter: data?.getCostCenterByIdStorefront,
+      role: permissionsData?.checkUserPermission?.role?.id ?? '',
+      roleName: permissionsData?.checkUserPermission?.role?.name ?? '',
     }
-  }, [data])
+  }, [data, permissionsData])
 
-  const loading = sessionLoading || organizationLoading
+  const loading = useMemo(
+    () => sessionLoading || organizationLoading || permissionsLoading,
+    [organizationLoading, permissionsLoading, sessionLoading]
+  )
 
   return { organization, loading }
 }