+++ react/hooks/usePermissions.ts
@@ -1,0 +1,17 @@
+import { useMemo } from 'react'
+
+import { useOrganization } from '.'
+
+export function usePermissions() {
+  const { organization } = useOrganization()
+  const { role } = organization
+
+  const isSalesUser = useMemo(
+    () =>
+      !!role &&
+      !['customer-admin', 'customer-approver', 'customer-buyer'].includes(role),
+    [role]
+  )
+
+  return { isSalesUser }
+}