+++ react/hooks/useTotalizers.tsx
@@ -1,24 +1,30 @@
-import React from 'react'
+import React, { useMemo } from 'react'
 import { useIntl } from 'react-intl'
 import { FormattedPrice } from 'vtex.formatted-price'
 import { IconHelp, Tooltip } from 'vtex.styleguide'
 
+import { useOrderFormCustom, useTotalMargin } from '.'
 import { PaymentData } from '../components/PaymentData'
 import { PONumber } from '../components/PONumber'
 import { TruncatedText } from '../components/TruncatedText'
-import { messages } from '../utils'
-import { useOrderFormCustom } from './useOrderFormCustom'
+import { B2B_QUOTES_CUSTOM_APP_ID, messages } from '../utils'
 
 export function useTotalizers() {
   const { formatMessage } = useIntl()
   const { orderForm } = useOrderFormCustom()
-  const { totalizers = [], value: total = 0, items } = orderForm
+  const { totalizers = [], customData, value: total = 0, items } = orderForm
+  const customApps = customData?.customApps
+  const hasQuotationDiscount = useMemo(
+    () =>
+      customApps?.find(
+        (app) => app.id === B2B_QUOTES_CUSTOM_APP_ID && app.fields?.quoteId
+      ),
+    [customApps]
+  )
+
+  const totalMargin = useTotalMargin()
 
   if (!totalizers.length || !items?.length) return []
-
-  const hasQuotationDiscount = items.some(
-    (item) => item.manualPrice && item.manualPrice < (item.price ?? 0)
-  )
 
   return [
     {
@@ -44,6 +50,16 @@ export function useTotalizers() {
         </div>
       ),
     })),
+    ...(totalMargin
+      ? [
+          {
+            label: formatMessage(messages.totalMargin),
+            value: (
+              <TruncatedText text={<FormattedPrice value={totalMargin} />} />
+            ),
+          },
+        ]
+      : []),
     {
       label: formatMessage(messages.total),
       value: <TruncatedText text={<FormattedPrice value={total / 100} />} />,