+++ react/typings/index.ts
@@ -1,5 +1,10 @@
 import type { Query } from 'vtex.b2b-organizations-graphql'
-import type { Address } from 'vtex.checkout-graphql'
+import type {
+  Address,
+  Maybe,
+  OrderForm as OrderFormType,
+} from 'vtex.checkout-graphql'
+import type { OrderForm as OrderFormStore } from 'vtex.store-graphql'
 
 type ShowToastArgs = {
   message: string
@@ -7,7 +12,7 @@ type ShowToastArgs = {
 }
 
 export type WithToast<T = unknown> = T & {
-  showToast?: (args: ShowToastArgs) => void
+  showToast: (args: ShowToastArgs) => void
 }
 
 export type TableSchema<RowType> = {
@@ -30,6 +35,25 @@ export type ApiResponse = {
   response?: { data?: { error?: string } | string }
   error?: { message?: string }
 }
+
+export type CompleteOrderFormData = ApiResponse &
+  Pick<OrderFormStore, 'sellers'> & {
+    orderFormId: string
+    invoiceData?: { address?: Maybe<Address> }
+    clientProfileData: OrderFormStore['clientProfileData'] & {
+      profileCompleteOnLoading?: string | null
+      profileErrorOnLoading?: string | null
+      customerClass?: string | null
+    }
+  }
+
+export type PaymentAddress = {
+  paymentAddress?: Maybe<Address>
+}
+
+export type CompleteOrderForm = OrderFormType &
+  PaymentAddress &
+  CompleteOrderFormData
 
 export type TransactionResponse = ApiResponse & {
   id: string
@@ -81,11 +105,14 @@ export type SessionOrganizationData = {
 
 export type GetOrganizationQuery = Pick<
   Query,
-  'getOrganizationByIdStorefront' | 'getUsers'
+  'getOrganizationByIdStorefront' | 'getUsers' | 'getCostCenterByIdStorefront'
 >
 
 export type CustomOrganization = GetOrganizationQuery['getOrganizationByIdStorefront'] & {
   users: GetOrganizationQuery['getUsers']
+  costCenter: GetOrganizationQuery['getCostCenterByIdStorefront']
+  role: string
+  roleName: string
 }
 
 type PaymentField<T> = {
@@ -95,4 +122,9 @@ type PaymentField<T> = {
 
 export type PaymentAddressType = {
   [K in keyof Address]: PaymentField<Address[K]>
+}
+
+export type ModalProps = {
+  open: boolean
+  setOpen: React.Dispatch<React.SetStateAction<boolean>>
 }